changes according to comments part-1
