+++ Service/src/main/java/SendingTheTelephone/TicketSending.java
@@ -3,5 +3,5 @@ package SendingTheTelephone;
 import TIcket.Ticket;
 
 public abstract class TicketSending {
-    public void shared(Ticket ticket){};
+    abstract public void shared(Ticket ticket);
 }
