+++ Service/src/main/java/org/example/SetClassId.java
@@ -1,6 +1,6 @@
 package org.example;
 
 public interface SetClassId {
-    void SetId(int id);
-    int GetId();
+    void setId(int id);
+    int getId();
 }
